% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{allTreeExposures}
\alias{allTreeExposures}
\title{Compute all tree exposures for different number of clusters}
\usage{
allTreeExposures(tree, feat_mat, sigs_filter, start = NULL, stop = NULL)
}
\arguments{
\item{tree}{graphNel}

\item{feat_mat}{data.frame}

\item{sigs_filter}{character vector}

\item{start}{numeric}

\item{stop}{numeric}
}
\value{
list of data.frame objects
}
\description{
This function takes as input a phylogeny, the signature vector, and the normalized feature matrix.
The phylogeny is represented with a graphNel object, and the normalized feature matrix has rows
corresponding to tree nodes and columns corresponding to trinucleotide contexts (e.g., "A(C > A)G").
It returns a list of exposure matrices whose entries correspond to the best partition of the
phylogeny for each possible number of clusters. Each exposure matrix has a row for each requested
signature and a column for each cluster of clones induced by the partition of the phylogeny.
Start and stop are optional parameters to specify the number of exposure shifts to try.
}
\examples{
allTreeExposures(T_tree, P_Norm, S, 0, 1)
}
