% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{E_list}
\alias{E_list}
\title{Sample list of exposure data frames}
\format{
A list of data frames, each with 2 rows (active signatures) and x variables (number of clone clusters):
\describe{E.g., B;C: Cluster of comprised of nodes B and C}
}
\usage{
E_list
}
\description{
A sample dataset containing a list mutational signature exposure dataframes for all possible numbers
of exposure shifts (i.e., k ranges from 0 to 8).
}
\keyword{datasets}
