% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandClusters}
\alias{expandClusters}
\title{Expand exposure dataframe width by spliting semicolons}
\usage{
expandClusters(d1)
}
\arguments{
\item{d1}{data.frame}
}
\value{
data.frame
}
\description{
Given a dataframe with column names containing semicolons,
splits column name at semicolon. The resulting string split generates new column names.
Each new column name is populated with a duplicate of its original column.
This is used to expand columns labeled by clusters of nodes into a column for each node.
}
\examples{
x <- data.frame(c(.2, .3, .5), c(.1, .7, .2))
colnames(x) <- c("a;b;c", "1;2;3")
expandClusters(x)
}
