% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastmerge}
\alias{fastmerge}
\title{Concatenate two dataframes}
\usage{
fastmerge(d1, d2)
}
\arguments{
\item{d1}{data.frame}

\item{d2}{data.frame}
}
\value{
data.frame
}
\description{
Create a new dataframe from two existing dataframes by stacking them. If, a column
exists in one dataframe but not the other, NA is added. This code was adapted
from https://stackoverflow.com/questions/8169323/r-concatenate-two-dataframes.
}
\examples{
fastmerge(data.frame("X" = 1:2, "Y" = c("a", "b")), data.frame("X" = 1:2, "Z" = c(5, 7)))
}
