% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{getBIC}
\alias{getBIC}
\title{Compute the Bayesian Information Criterion}
\usage{
getBIC(feat_mat, exp_mat, sigs_filter)
}
\arguments{
\item{feat_mat}{data.frame}

\item{exp_mat}{data.frame}

\item{sigs_filter}{character vector}
}
\value{
numeric
}
\description{
Given a normalized feature matrix, estimated exposures, and signature matrix, this function returns
the Bayesian information criterion based on the error between the reconstructed feature matrix and
the normalized feature matrix. In the feature matrix, each row is a clone and each column is a
trinucleotide context. In the exposure matrix, each row corresponds to a mutational signature and
each column corresponds to a cluster of clones. Finally, in the signature vector, the signatures of
interest are given (e.g., "Signature.1").
}
\examples{
getBIC(P_Norm, E_list[[1]], S)
}
