% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{normalizeFeatureMatrix}
\alias{normalizeFeatureMatrix}
\title{Normalize mutational signature feature matrix}
\usage{
normalizeFeatureMatrix(feat_mat, norm_method)
}
\arguments{
\item{feat_mat}{data.frame}

\item{norm_method}{character}
}
\value{
data.frame
}
\description{
This function takes as input feature matrix where each row is a clone and each
column is a trinucleotide context (e.g., "A(C > A)G"). It then adjusts the counts
based on the trinucleotide context fraction. The backend of this function relies
getTriContextFraction in deconstuctSigs. Possible normalization options include:
on "default", "genome", and "exome". Please see deconstructSigs documentation
for more information on the normalization procedure.
}
\examples{
tri.counts.genome <- deconstructSigs::tri.counts.genome
normalizeFeatureMatrix(P, "genome")
}
