% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{outputExposures}
\alias{outputExposures}
\title{Format exposure output data frame for given number of exposure shifts}
\usage{
outputExposures(sample, tree_idx, exp_list, k)
}
\arguments{
\item{sample}{character}

\item{tree_idx}{numeric}

\item{exp_list}{list of data.frame objects}

\item{k}{numeric}
}
\value{
data.frame
}
\description{
Given a sample ID (e.g., patient tumor ID), tree ID for this sample (e.g., 2), list of exposure
data frames for all numbers of exposure shifts, and the number of exposure shifts of interest,
this function returned a dataframe where each row corresponds to a node in the sample tree. The
columns are as follows: Sample ID, Tree ID, k, Node ID, Signature 1 exposure ... Signature X exposure.
This formatting is used in the input to the corresponding visualization tool for PhySigs.
}
\examples{
outputExposures("Tumor1", 0, E_list, 1)
}
