% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{outputTrees}
\alias{outputTrees}
\title{Format tree output data frame}
\usage{
outputTrees(sample, tree_idx, tree)
}
\arguments{
\item{sample}{character}

\item{tree_idx}{numeric}

\item{tree}{graphNel}
}
\value{
data.frame
}
\description{
Given a sample ID (e.g., patient tumor ID), tree ID for this sample (e.g., 2), and a tree object,
This function returns a dataframe with one row corresponding to the input tree. The columns are
as follows: Sample ID; Tree ID; Number of clones/nodes; Nodes; Edges (source -> target).
This formatting is used in the input to the corresponding visualization tool for PhySigs.
}
\examples{
outputTrees("Tumor1", 0, T_tree)
}
