% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P}
\alias{P}
\title{Sample raw feature matrix}
\format{
A data frame with 9 rows and 96 variables:
\describe{A[C>A]A: trinucleotide context count, C to A substitution flanked by A and A}
}
\usage{
P
}
\description{
A sample dataset containing the trinucleotide counts for each node (clone) in
tumor phylogeny T.
}
\keyword{datasets}
