% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{Plot tree with exposure pie chart at each node}
\usage{
plotTree(patient, title, tree, exp_mat, tree_idx = 0)
}
\arguments{
\item{patient}{character}

\item{title}{character}

\item{tree}{graphNel}

\item{exp_mat}{data.frame}

\item{tree_idx}{numeric}
}
\description{
Given a tree represented by a graphNel object, this function prints the tree with each
node labeled by a pie chart showing the exposure to mutational signatures at that node.
Patient, title, and tree_idx fields just label the figure. The exposure matrix should have
a row for each active mutational signature and a column for each clone cluster.
}
\examples{
plotTree("Tumor1", "Example Plot", T_tree, E_list[[1]], tree_idx=1)
}
