% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S}
\alias{S}
\title{Sample set of active mutational signatures.}
\format{
A character vector
}
\usage{
S
}
\description{
A sample character vector containing subset of active mutational signatures to include.
}
\keyword{datasets}
