% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T_tree}
\alias{T_tree}
\title{Sample tumor phylogeny.}
\format{
A graphNEL object with 9 nodes:
\describe{
\item{nodes}{clone, with node ID label}
}
}
\usage{
T_tree
}
\description{
A sample graphNEL tree where the nodes correspond to clones.
}
\keyword{datasets}
