% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{treeExposures}
\alias{treeExposures}
\title{Compute best exposure for k exposure shifts}
\usage{
treeExposures(tree, feat_mat, k, sigs_filter)
}
\arguments{
\item{tree}{graphNel}

\item{feat_mat}{data.frame}

\item{k}{numeric}

\item{sigs_filter}{character vector}
}
\value{
data.frame
}
\description{
This function takes as input a phylogeny, the signature vector, and the normalized feature matrix.
It also takes as input k, the desired number of exposure shifts to be found in the input phylogeny.
The phylogeny is represented with a graphNel object, and the normalized feature matrix has rows
corresponding to tree nodes and columns corresponding to trinucleotide contexts (e.g., "A(C > A)G").
It returns a list of exposure matrices whose entries correspond to the best partition of the
phylogeny for each possible number of clusters. Each exposure matrix has a row for each requested
signature and a column for each cluster of clones induced by the partition of the phylogeny.
}
\examples{
treeExposures(T_tree, P_Norm, 1, S)
}
