% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tri.counts.exome}
\alias{tri.counts.exome}
\title{The counts of every trinuclotide frequency in an exome}
\format{
A data frame of 32 rows and 1 column that contains the counts
}
\source{
\url{https://rdrr.io/cran/deconstructSigs/man/tri.counts.exome.html}
}
\usage{
tri.counts.exome
}
\description{
A datset containing the number of times each trinucleotide (ex: ACA) is found in the exome region captured
by sequencing.
}
\keyword{datasets}
